Move GenerateCommandOutputInterface
