+++ src/Model/GenerateCommand/AbstractOutput.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace webignition\BasilRunner\Model\GenerateCommand;
 
 use webignition\BasilRunner\Model\GenerateCommandConfiguration;
-use webignition\BasilRunner\Model\GenerateCommandOutputInterface;
 
-abstract class AbstractOutput implements GenerateCommandOutputInterface
+abstract class AbstractOutput implements OutputInterface
 {
     protected const STATUS_SUCCESS = 'success';
     protected const STATUS_FAILURE = 'failure';
